cambios

+++ app/Http/Controllers/AppController.php
@@ -117,8 +117,15 @@ class AppController extends Controller
         }
 
 
-        public function fillTableCodes()
+        public function fillTableCodes(Request $request)
         {
+
+            $allowIp = ['10.124.0.9'];
+
+            if(!in_array($request->ip(),$allowIp))
+            {
+                return response()->json(['error' => 'No tienes permiso para acceder a esta ruta'],404);
+            }
             CodeAccess::truncate(); // Elimina todos los registros existentes
         
             $users = User::where('rol_id', 1)->get();
@@ -144,6 +151,8 @@ class AppController extends Controller
             }
             $allCodes = CodeAccess::all();
             return response()->json(['success' => 'Tabla llenada con éxito', 'codes' => $allCodes], 200);
+      
+           
         }
 
         public function logout(Request $request)
